services/horizon: Move toid package to root directory (#4215)
Move `toid` package to root directory from Horizon's `internal` package.

`toid` can be useful in other services. It was also made a standard SEP:
https://github.com/stellar/stellar-protocol/blob/master/ecosystem/sep-0035.md
