+++ services/horizon/internal/db2/history/effect_batch_insert_builder_test.go
@@ -6,7 +6,7 @@ import (
 
 	"github.com/guregu/null"
 	"github.com/stellar/go/services/horizon/internal/test"
-	"github.com/stellar/go/services/horizon/internal/toid"
+	"github.com/stellar/go/toid"
 )
 
 func TestAddEffect(t *testing.T) {
