+++ services/horizon/internal/db2/history/ledger_test.go
@@ -12,7 +12,7 @@ import (
 	"github.com/guregu/null"
 	"github.com/stellar/go/ingest/ledgerbackend"
 	"github.com/stellar/go/services/horizon/internal/test"
-	"github.com/stellar/go/services/horizon/internal/toid"
+	"github.com/stellar/go/toid"
 	"github.com/stellar/go/xdr"
 )
