+++ services/horizon/internal/db2/history/trade_aggregation.go
@@ -8,9 +8,9 @@ import (
 	sq "github.com/Masterminds/squirrel"
 
 	"github.com/stellar/go/services/horizon/internal/db2"
-	"github.com/stellar/go/services/horizon/internal/toid"
 	"github.com/stellar/go/support/errors"
 	strtime "github.com/stellar/go/support/time"
+	"github.com/stellar/go/toid"
 )
 
 // AllowedResolutions is the set of trade aggregation time windows allowed to be used as the
