+++ services/horizon/internal/db2/history/trade_scenario.go
@@ -8,7 +8,7 @@ import (
 	"github.com/guregu/null"
 
 	"github.com/stellar/go/services/horizon/internal/test"
-	"github.com/stellar/go/services/horizon/internal/toid"
+	"github.com/stellar/go/toid"
 	"github.com/stellar/go/xdr"
 )
