+++ services/horizon/internal/db2/history/transaction_batch_insert_builder.go
@@ -15,10 +15,10 @@ import (
 	"github.com/lib/pq"
 
 	"github.com/stellar/go/ingest"
-	"github.com/stellar/go/services/horizon/internal/toid"
 	"github.com/stellar/go/services/horizon/internal/utf8"
 	"github.com/stellar/go/support/db"
 	"github.com/stellar/go/support/errors"
+	"github.com/stellar/go/toid"
 	"github.com/stellar/go/xdr"
 )
