+++ services/horizon/internal/db2/history/transaction_test.go
@@ -11,7 +11,7 @@ import (
 
 	"github.com/stellar/go/ingest"
 	"github.com/stellar/go/services/horizon/internal/test"
-	"github.com/stellar/go/services/horizon/internal/toid"
+	"github.com/stellar/go/toid"
 )
 
 func TestTransactionQueries(t *testing.T) {
